# TraceService/src/WebAPI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TraceService",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/traces/Enquiry": {
      "post": {
        "tags": [
          "TraceEntities"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "string"
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BancontactTraceModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BancontactTraceModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BancontactTraceModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traces/Search": {
      "post": {
        "tags": [
          "TraceEntities"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/SearchModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SearchModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedList`1"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedList`1"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedList`1"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BancontactReceivedModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dt": {
            "type": "string",
            "format": "date-time"
          },
          "mti": {
            "type": "string",
            "nullable": true
          },
          "trn": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BancontactSentModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dt": {
            "type": "string",
            "format": "date-time"
          },
          "mti": {
            "type": "string",
            "nullable": true
          },
          "trn": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BancontactTraceModel": {
        "type": "object",
        "properties": {
          "bancontactReceived": {
            "$ref": "#/components/schemas/BancontactReceivedModel"
          },
          "bancontactSent": {
            "$ref": "#/components/schemas/BancontactSentModel"
          },
          "cslInBsad": {
            "$ref": "#/components/schemas/CslInBsadModel"
          },
          "cslInBsau": {
            "$ref": "#/components/schemas/CslInBsauModel"
          },
          "cslOutBsad": {
            "$ref": "#/components/schemas/CslOutBsadModel"
          },
          "cslOutBsau": {
            "$ref": "#/components/schemas/CslOutBsauModel"
          }
        },
        "additionalProperties": false
      },
      "CslInBsadModel": {
        "type": "object",
        "properties": {
          "mid": {
            "type": "string",
            "nullable": true
          },
          "tid": {
            "type": "string",
            "nullable": true
          },
          "rrnIn": {
            "type": "string",
            "nullable": true
          },
          "rrnOut": {
            "type": "string",
            "nullable": true
          },
          "procCode": {
            "type": "string",
            "nullable": true
          },
          "panExpDate": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dt": {
            "type": "string",
            "format": "date-time"
          },
          "mti": {
            "type": "string",
            "nullable": true
          },
          "trn": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CslInBsauModel": {
        "type": "object",
        "properties": {
          "mid": {
            "type": "string",
            "nullable": true
          },
          "tid": {
            "type": "string",
            "nullable": true
          },
          "rrnIn": {
            "type": "string",
            "nullable": true
          },
          "rrnOut": {
            "type": "string",
            "nullable": true
          },
          "procCode": {
            "type": "string",
            "nullable": true
          },
          "panExpDate": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dt": {
            "type": "string",
            "format": "date-time"
          },
          "mti": {
            "type": "string",
            "nullable": true
          },
          "trn": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CslOutBsadModel": {
        "type": "object",
        "properties": {
          "rrnCardScheme": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dt": {
            "type": "string",
            "format": "date-time"
          },
          "mti": {
            "type": "string",
            "nullable": true
          },
          "trn": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CslOutBsauModel": {
        "type": "object",
        "properties": {
          "rrnCardScheme": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dt": {
            "type": "string",
            "format": "date-time"
          },
          "mti": {
            "type": "string",
            "nullable": true
          },
          "trn": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PaginatedList`1": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResultModel"
            },
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "readOnly": true
          },
          "hasNextPage": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "SearchModel": {
        "type": "object",
        "properties": {
          "dateFrom": {
            "type": "string",
            "format": "date-time"
          },
          "dateTo": {
            "type": "string",
            "format": "date-time"
          },
          "mid": {
            "type": "string",
            "nullable": true
          },
          "tid": {
            "type": "string",
            "nullable": true
          },
          "pan": {
            "type": "string",
            "nullable": true
          },
          "expDate": {
            "type": "string",
            "nullable": true
          },
          "rrn": {
            "type": "string",
            "nullable": true
          },
          "mti": {
            "type": "string",
            "nullable": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SearchResultModel": {
        "type": "object",
        "properties": {
          "trn": {
            "type": "string",
            "nullable": true
          },
          "mti": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Please insert Bearer Token into field",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}